Moves license to standard location
GitHub can hook into it, various license harvesting tools have a nicer
time auditing whether licenses used in a given project are compatible if
it sees a files named similar to `LICENSE{,.txt,.md}` in the base of the
repository.

No changes made to the wording, but shifted it to a more standard
location.